# openapi.yaml
openapi: 3.0.3
info:
  title: Athletic AI API
  description: API para acompanhamento de treinos e nutrição com IA
  version: 1.0.0
  contact:
    name: Athletic AI Team
servers:
  - url: http://localhost:8000/api/v1
    description: Desenvolvimento local
  - url: https://api.athletic.ai/api/v1
    description: Produção
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WorkoutCreate:
      type: object
      required: [exercise, muscle_group, sets, reps, weight_kg, workout_date]
      properties:
        exercise: { type: string }
        muscle_group: { type: string }
        sets: { type: integer, minimum: 1 }
        reps: { type: integer, minimum: 1 }
        weight_kg: { type: number, minimum: 0 }
        rpe: { type: integer, minimum: 1, maximum: 10 }
        notes: { type: string }
        workout_date: { type: string, format: date-time }
    WorkoutResponse:
      allOf:
        - $ref: "#/components/schemas/WorkoutCreate"
        - type: object
          properties:
            id: { type: string, format: uuid }
            user_id: { type: string, format: uuid }
            created_at: { type: string, format: date-time }
...
paths:
...
  /workouts:
    get:
      tags: [workouts]
      summary: Listar treinos
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20 }
        - name: muscle_group
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Lista paginada de treinos
    post:
      tags: [workouts]
      summary: Criar treino
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkoutCreate" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkoutResponse" }
  /workouts/{workout_id}:
    get:
      tags: [workouts]
      summary: Buscar treino por ID
      parameters:
        - name: workout_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkoutResponse" }
        "404":
          description: Não encontrado
    put:
      tags: [workouts]
      summary: Atualizar treino
      parameters:
        - name: workout_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkoutResponse" }
    delete:
      tags: [workouts]
      summary: Remover treino
      parameters:
        - name: workout_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Removido com sucesso
...
  /photos:
    get:
      tags: [photos]
      summary: Listar fotos corporais
      responses:
        "200":
          description: Lista de fotos
    post:
      tags: [photos]
      summary: Upload de foto corporal
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                weight_kg: { type: number }
                notes: { type: string }
                taken_at: { type: string, format: date-time }
      responses:
        "201":
          description: Foto enviada
  /photos/{photo_id}:
    delete:
      tags: [photos]
      summary: Remover foto
      parameters:
        - name: photo_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Removida
...